requirements.txt
pysnmp==4.4.12
paho-mqtt==1.6.1
